refactor(cli): split tsserver completions helpers
