+++ crates/tsz-cli/src/bin/tsz_server/main.rs
@@ -39,8 +39,10 @@ mod handlers_code_fixes_jsdoc;
 mod handlers_code_fixes_synthetic;
 mod handlers_code_fixes_utils;
 mod handlers_completions;
+mod handlers_completions_auto_imports;
 mod handlers_completions_display;
 mod handlers_completions_parameters;
+mod handlers_completions_project;
 mod handlers_completions_snippets;
 mod handlers_diagnostics;
 mod handlers_editing;
